fondos_app/requirements.txt
streamlit
numpy
pandas
polars
fastexcel
pillow
python-dotenv
plotly
great-tables
